Moved MvcUriBuilder class to javax.mvc.uri package
